Visualizer: Adjust for volume level
Volume levels are taken into account when attaching to sessionId=0
(system mixer output). Account for this by querying volume and
multiplying the reciprocal of the percentage. Unfortunately, there is
nothing that can be done about volume=0

We could refactor to attach to specific audio sessions, but this limits
us to one stream only and adds complexity

Change-Id: I7ccf2852a5018a72eabf0c7dbdd2f7242eac58ce